+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Add unit tests for Event:
   - voucherCreated
   - AccountAuthorized, AccountUnauthorized (#20)
+  - VoucherTypeDurationUpdated, VoucherTypeDescriptionUpdated (#21)
   - VoucherTypeCreated (#22)
 
 ## v1.0.0